# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RiverKin API",
    "description": "RiverKin backend \u2014 IEEE OneAquaHealth Global Hackathon 2026 (Track 5).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/observations": {
      "post": {
        "tags": [
          "observations"
        ],
        "summary": "Create Observation",
        "description": "Submit a field check: safety + quality gates, then draft verify items.\n\nSafety refusals return 403 + ``safety_reason`` (PRD). A pipe/outfall answer\nroutes the observation straight to expert review.",
        "operationId": "create_observation_api_v1_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/observations/{observation_id}/photos": {
      "post": {
        "tags": [
          "observations"
        ],
        "summary": "Upload Photo",
        "description": "Upload one observation photo.\n\nRuns EXIF strip + Laplacian blur score + face blur + pHash, rejects reused\nimages (pHash distance < 8 against the site's past photos), and stores the\nprocessed bytes. A blurry photo returns 422 ``retake_photo``.",
        "operationId": "upload_photo_api_v1_observations__observation_id__photos_post",
        "parameters": [
          {
            "name": "observation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Observation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photo_api_v1_observations__observation_id__photos_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Photo Api V1 Observations  Observation Id  Photos Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/observations/{observation_id}/status": {
      "get": {
        "tags": [
          "observations"
        ],
        "summary": "Observation Status",
        "description": "Live receipt + verification status for one observation.",
        "operationId": "observation_status_api_v1_observations__observation_id__status_get",
        "parameters": [
          {
            "name": "observation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Observation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fhir/observations/{observation_id}": {
      "get": {
        "tags": [
          "fhir"
        ],
        "summary": "Get Fhir Observation",
        "description": "Return the FHIR Bundle for one observation (optionally POST to HAPI).",
        "operationId": "get_fhir_observation_api_v1_fhir_observations__observation_id__get",
        "parameters": [
          {
            "name": "observation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Observation Id"
            }
          },
          {
            "name": "post",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Post"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Fhir Observation Api V1 Fhir Observations  Observation Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_photo_api_v1_observations__observation_id__photos_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "upstream"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_photo_api_v1_observations__observation_id__photos_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObservationCreated": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verify_item_count": {
            "type": "integer",
            "title": "Verify Item Count"
          },
          "receipt": {
            "$ref": "#/components/schemas/ReceiptOut"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "verify_item_count",
          "receipt"
        ],
        "title": "ObservationCreated",
        "description": "Response to POST /observations \u2014 the new id + first receipt."
      },
      "ObservationIn": {
        "properties": {
          "site_code": {
            "type": "string",
            "title": "Site Code"
          },
          "answers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Answers",
            "default": {}
          },
          "feeling": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feeling"
          },
          "photo_count": {
            "type": "integer",
            "title": "Photo Count",
            "default": 0
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          }
        },
        "type": "object",
        "required": [
          "site_code"
        ],
        "title": "ObservationIn",
        "description": "Field-check submission from C4.\n\n``answers`` is keyed by the front-end question id (``q-water`` \u2026), which\nmaps to the OAH field codes in app/ai/questions.py::FIELD_SPECS. ``lat``/\n``lng`` are used once for the geofence check and then discarded \u2014 raw GPS is\nnever persisted (PRD privacy rule)."
      },
      "ObservationStatusOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trust": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust"
          },
          "verifier_count": {
            "type": "integer",
            "title": "Verifier Count"
          },
          "receipt": {
            "$ref": "#/components/schemas/ReceiptOut"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "trust",
          "verifier_count",
          "receipt"
        ],
        "title": "ObservationStatusOut",
        "description": "GET /observations/{id}/status \u2014 live trust + receipt."
      },
      "ReceiptOut": {
        "properties": {
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "waterbody": {
            "type": "string",
            "title": "Waterbody"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "gap_before": {
            "type": "integer",
            "title": "Gap Before"
          },
          "gap_after": {
            "type": "integer",
            "title": "Gap After"
          },
          "rain_context": {
            "type": "string",
            "title": "Rain Context"
          },
          "verifier_count": {
            "type": "integer",
            "title": "Verifier Count"
          },
          "fhir_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fhir Id"
          },
          "sentinel_line": {
            "type": "string",
            "title": "Sentinel Line"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "date_label": {
            "type": "string",
            "title": "Date Label"
          }
        },
        "type": "object",
        "required": [
          "site_name",
          "waterbody",
          "city",
          "gap_before",
          "gap_after",
          "rain_context",
          "verifier_count",
          "fhir_id",
          "sentinel_line",
          "state",
          "date_label"
        ],
        "title": "ReceiptOut",
        "description": "Impact receipt rendered on C6."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}